+++ crates/external-data/src/poscar.rs
@@ -6,6 +6,8 @@ use vasp_poscar::Poscar;
 
 use velvet_core::prelude::*;
 
+use crate::internal::Float;
+
 /// Returns a [`System`](velvet_core::system::System) object initialized from POSCAR formatted data.
 ///
 /// # Arguments
@@ -47,16 +49,16 @@ where
 
     // Set system cell.
     let vecs = poscar.scaled_lattice_vectors();
-    let matrix: Matrix3<f64> = Matrix3::new(
-        vecs[0][0] as f64,
-        vecs[1][0] as f64,
-        vecs[2][0] as f64,
-        vecs[0][1] as f64,
-        vecs[1][1] as f64,
-        vecs[2][1] as f64,
-        vecs[0][2] as f64,
-        vecs[1][2] as f64,
-        vecs[2][2] as f64,
+    let matrix: Matrix3<Float> = Matrix3::new(
+        vecs[0][0] as Float,
+        vecs[1][0] as Float,
+        vecs[2][0] as Float,
+        vecs[0][1] as Float,
+        vecs[1][1] as Float,
+        vecs[2][1] as Float,
+        vecs[0][2] as Float,
+        vecs[1][2] as Float,
+        vecs[2][2] as Float,
     );
     let cell = Cell::from_matrix(matrix);
     builder = builder.with_cell(cell);
@@ -74,18 +76,18 @@ where
     }
 
     // Set system positions.
-    let positions: Vec<Vector3<f64>> = poscar
+    let positions: Vec<Vector3<Float>> = poscar
         .scaled_cart_positions()
         .iter()
-        .map(|x| Vector3::new(x[0] as f64, x[1] as f64, x[2] as f64))
+        .map(|x| Vector3::new(x[0] as Float, x[1] as Float, x[2] as Float))
         .collect();
     builder = builder.with_positions(positions);
 
     // Set system velocities if they exist.
     if let Some(vels) = poscar.cart_velocities() {
-        let velocities: Vec<Vector3<f64>> = vels
+        let velocities: Vec<Vector3<Float>> = vels
             .iter()
-            .map(|x| Vector3::new(x[0] as f64, x[1] as f64, x[2] as f64))
+            .map(|x| Vector3::new(x[0] as Float, x[1] as Float, x[2] as Float))
             .collect();
         builder = builder.with_velocities(velocities);
     }
